# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/suppliers": {
      "post": {
        "description": "Permite crear un nuevo proveedor de repuestos",
        "operationId": "SuppliersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuppliersCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proveedor creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppliersCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "409": {
            "description": "El recurso ya existe"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Crear nuevo proveedor",
        "tags": [
          "Suppliers"
        ]
      },
      "get": {
        "description": "Permite obtener una lista paginada de proveedores",
        "operationId": "SuppliersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Proveedores encontrados correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Listar proveedores",
        "tags": [
          "Suppliers"
        ]
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "description": "Permite obtener un proveedor específico con sus productos",
        "operationId": "SuppliersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor encontrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppliersCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener proveedor por ID",
        "tags": [
          "Suppliers"
        ]
      },
      "patch": {
        "description": "Permite actualizar un proveedor existente",
        "operationId": "SuppliersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuppliersUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppliersUpdateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "409": {
            "description": "El recurso ya existe"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Actualizar proveedor",
        "tags": [
          "Suppliers"
        ]
      },
      "delete": {
        "description": "Permite eliminar un proveedor (soft delete) si no tiene productos activos",
        "operationId": "SuppliersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor eliminado correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Eliminar proveedor",
        "tags": [
          "Suppliers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Almacén API",
    "description": "API para gestión de almacén - Productos, Cotizaciones, Ventas y Proveedores",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SuppliersCreateRequestDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Auto Parts Inc."
          },
          "telefono": {
            "type": "string",
            "example": "+1234567890"
          },
          "email": {
            "type": "string",
            "example": "contacto@autoparts.com"
          },
          "direccion": {
            "type": "string",
            "example": "Av. Principal 123"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "SuppliersCreateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Supplier created successfully"
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "example": "Auto Parts Inc."
          },
          "telefono": {
            "type": "object",
            "example": "+1234567890"
          },
          "email": {
            "type": "object",
            "example": "contacto@autoparts.com"
          },
          "direccion": {
            "type": "object",
            "example": "Av. Principal 123"
          },
          "activo": {
            "type": "boolean",
            "example": true
          },
          "fechaRegistro": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "status_code": {
            "type": "number",
            "example": 201
          }
        },
        "required": [
          "message",
          "id",
          "nombre",
          "activo",
          "fechaRegistro",
          "status_code"
        ]
      },
      "SuppliersUpdateRequestDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Auto Parts Inc. Updated"
          },
          "telefono": {
            "type": "string",
            "example": "+1234567890"
          },
          "email": {
            "type": "string",
            "example": "contacto@autoparts.com"
          },
          "direccion": {
            "type": "string",
            "example": "Av. Principal 456"
          }
        }
      },
      "SuppliersUpdateResponseDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}